Ajustes de chore e session

+++ Repository/Corp/Entities/ChoreEntity.cs
@@ -2,12 +2,12 @@
 {
     public partial class ChoreEntity
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
         public bool IsCompleted { get; set; }
         public DateTime CreatedAt { get; set; }
-        public int SessionId { get; set; }
+        public long SessionId { get; set; }
         public virtual SessionEntity Session { get; set; }
     }
 }
